Extract audit diagnostic details builder
